+++ components/command.js
@@ -12,6 +12,7 @@ import {
   CommandShortcut,
 } from "@/components/ui/command"
 import Link from "next/link"
+import { HomeIcon, FileTextIcon, LayersIcon, CalendarIcon, IdCardIcon, ReaderIcon, InfoCircledIcon, QuestionMarkCircledIcon, BlendingModeIcon, EyeOpenIcon, LaptopIcon, FileIcon, RocketIcon } from "@radix-ui/react-icons"
 
 export function Commands() {
   const [open, setOpen] = React.useState(false)
@@ -35,27 +36,98 @@ export function Commands() {
         <CommandSeparator />
         <CommandGroup heading="Main pages">
           <CommandSeparator />
-          <Link href={"/"}><CommandItem>Home</CommandItem></Link>
-          <Link href={"/docs"}><CommandItem>Docs</CommandItem></Link>
-          <CommandItem>Templates</CommandItem>
+
+          <Link href={"/"}>
+            <CommandItem>
+              <HomeIcon className="mr-2 h-4 w-4" />
+              <span className="font-inter">Home</span>
+            </CommandItem>
+          </Link>
+
+          <Link href={"/docs"}>
+            <CommandItem>
+              <FileTextIcon className="mr-2 h-4 w-4" />
+              <span className="font-inter">Documentation</span>
+            </CommandItem>
+          </Link>
+
+          <Link href={"/"}>
+            <CommandItem>
+              <LayersIcon className="mr-2 h-4 w-4" />
+              <span className="font-inter">Templates</span>
+            </CommandItem>
+          </Link>
+
         </CommandGroup>
         <CommandSeparator />
+        
         <CommandGroup heading="Templates">
           <CommandSeparator />
-          <Link href={"/dashboard"}><CommandItem>Dashboard</CommandItem></Link>
-          <Link href={"/pricing-page"}><CommandItem>Pricing Page</CommandItem></Link>
-          <Link href={"/blog-post"}><CommandItem>Blog Post</CommandItem></Link>
+          <Link href={"/dashboard"}>
+            <CommandItem>
+              <CalendarIcon className="mr-2 h-4 w-4" />
+              <span className="font-inter">Dashboard</span>
+            </CommandItem>
+          </Link>
+          <Link href={"/pricing-page"}>
+            <CommandItem>
+              <IdCardIcon className="mr-2 h-4 w-4" />
+              <span className="font-inter">Pricing page</span>
+            </CommandItem>
+          </Link>
+          <Link href={"/blog-post"}>
+            <CommandItem>
+              <ReaderIcon className="mr-2 h-4 w-4" />
+              <span className="font-inter">Blog post</span>
+            </CommandItem>
+          </Link>
         </CommandGroup>
+
         <CommandSeparator />
         <CommandGroup heading="Documentation">
           <CommandSeparator />
-          <Link href={"/docs#introduction"}><CommandItem>Introduction</CommandItem></Link>
-          <Link href={"/docs#how-it-works"}><CommandItem>How does it work?</CommandItem></Link>
-          <Link href={"/docs#colors"}><CommandItem>Selecting Colors</CommandItem></Link>
-          <Link href={"/docs#theming"}><CommandItem>Theming</CommandItem></Link>
-          <Link href={"/docs#Upcoming"}><CommandItem>Upcoming</CommandItem></Link>
-          <Link href={"/docs#changelog"}><CommandItem>Changelog</CommandItem></Link>
-          <Link href={"/docs#version"}><CommandItem>Version</CommandItem></Link>
+          <Link href={"/docs#introduction"}>
+            <CommandItem>
+              <InfoCircledIcon className="mr-2 h-4 w-4" />
+              <span className="font-inter">Introduction</span>
+            </CommandItem>
+          </Link>
+          <Link href={"/docs#how-it-works"}>
+            <CommandItem>
+              <QuestionMarkCircledIcon className="mr-2 h-4 w-4" />
+              <span className="font-inter">How does it work?</span>
+            </CommandItem>
+          </Link>
+          <Link href={"/docs#colors"}>
+            <CommandItem>
+              <BlendingModeIcon className="mr-2 h-4 w-4" />
+              <span className="font-inter">Selecting colors</span>
+            </CommandItem>
+          </Link>
+          <Link href={"/docs#theming"}>
+            <CommandItem>
+              <EyeOpenIcon className="mr-2 h-4 w-4" />
+              <span className="font-inter">Theming</span>
+            </CommandItem>
+          </Link>
+          <Link href={"/docs#upcoming"}>
+            <CommandItem>
+              <LaptopIcon className="mr-2 h-4 w-4" />
+              <span className="font-inter">Upcoming</span>
+            </CommandItem>
+          </Link>
+          <Link href={"/docs#changelog"}>
+            <CommandItem>
+              <FileIcon className="mr-2 h-4 w-4" />
+              <span className="font-inter">Changelog</span>
+            </CommandItem>
+          </Link>
+          <Link href={"/docs#version"}>
+            <CommandItem>
+              <RocketIcon className="mr-2 h-4 w-4" />
+              <span className="font-inter">Version</span>
+            </CommandItem>
+          </Link>
         </CommandGroup>
       </CommandList>
     </CommandDialog>